requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
apscheduler
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
